refactor: Title 컴포넌트 이동
