+++ app/note/page.tsx
@@ -2,9 +2,9 @@ import { getCategories } from "@/notion/server/getCategories";
 import { getNotes } from "@/notion/server/getNotes";
 
 import { Divider } from "../../component/common/Divider";
+import { Title } from "../../component/common/Title";
 import { CategoryNav } from "../../component/note/CategoryNav";
 import { NoteList } from "../../component/note/NoteList";
-import { Title } from "../../component/Title";
 
 const Page = async () => {
   const categorys = await getCategories();
